refactor: remove file index and no-header and rename active-cloumn to index
