+++ src/views/EdgeNode/ListViewTabServices.vue
@@ -2,7 +2,7 @@
   import Illustration from '@/assets/svg/illustration-layers'
   import EmptyResultsBlock from '@/templates/empty-results-block'
   import { columnBuilder } from '@/templates/list-table-block/columns/column-builder'
-  import ListTableBlock from '@/templates/list-table-block/action-column.vue'
+  import ListTableBlock from '@/templates/list-table-block/index.vue'
   import DrawerService from '@/views/EdgeNode/Drawer'
   import PrimeButton from 'primevue/button'
   import { computed, ref } from 'vue'
