+++ src/views/NetworkLists/ListView.vue
@@ -36,7 +36,7 @@
   import Illustration from '@/assets/svg/illustration-layers.vue'
   import ContentBlock from '@/templates/content-block'
   import EmptyResultsBlock from '@/templates/empty-results-block'
-  import ListTableBlock from '@/templates/list-table-block/action-column.vue'
+  import ListTableBlock from '@/templates/list-table-block/index.vue'
   import PageHeadingBlock from '@/templates/page-heading-block'
 
   defineOptions({ name: 'network-list-view' })
