+++ src/views/RealTimeEventsDataStream/ListView.vue
@@ -1,6 +1,6 @@
 <script setup>
   import EmptyResultsBlock from '@/templates/empty-results-block'
-  import ListTableBlock from '@/templates/list-table-block/action-column.vue'
+  import ListTableBlock from '@/templates/list-table-block/index.vue'
   import Drawer from './Drawer'
   import PrimeButton from 'primevue/button'
   import { computed, ref } from 'vue'
