+++ src/views/WafRules/ListView.vue
@@ -33,7 +33,7 @@
 <script setup>
   import { ref, computed } from 'vue'
 
-  import ListTableBlock from '@/templates/list-table-block/action-column.vue'
+  import ListTableBlock from '@/templates/list-table-block/index.vue'
   import EmptyResultsBlock from '@/templates/empty-results-block'
   import Illustration from '@/assets/svg/illustration-layers.vue'
   import ContentBlock from '@/templates/content-block'
